feat: Update plan structures to DownUpOrder<f32> for accurate saturation
CRITICAL MONITORING ACCURACY UPDATE:
- Update IpStatsWithPlan.plan: DownUpOrder<u32> → DownUpOrder<f32>
- Update Circuit.plan: DownUpOrder<u32> → DownUpOrder<f32>
- Remove rate_for_plan() conversion functions (no longer needed)
- Use struct literal syntax to avoid trait bound issues with f32

IMPACT:
- Saturation calculations now use precise fractional rates
- Top N and Worst N widgets show accurate utilization percentages
- Example: 2.3/2.5 Mbps = 92% (was 77% with rounding)
- Network congestion detection restored to full accuracy
- LTS/Insight compatibility maintained via rate_for_submission()

All tests pass, system compiles successfully.

+++ src/rust/lqos_bus/src/ip_stats.rs
@@ -208,7 +208,7 @@ pub struct Circuit {
     /// The device name
     pub device_name: Option<String>,
     /// The current plan for this circuit.
-    pub plan: DownUpOrder<u32>,
+    pub plan: DownUpOrder<f32>,
     /// The last time this host was seen, in nanoseconds since boot.
     pub last_seen_nanos: u64,
 }
